feat: correcoes PDV e eventos de exclusao
- PDV: NPE ao finalizar venda sem cliente resolvido (usa CLIENTE PADRAO)
- PDV: dataCriacao adicionado aos itens do pedido
- PedidoModel.clienteId alterado de Long para Integer (Persism)
- ContaAreceberService: dataCriacao adicionado ao gerar contas
- VendaMercadoriaScreen: evento DadosFinanceirosAtualizados ao excluir
- ComprasScreen: evento DadosFinanceirosAtualizados ao excluir
- PDVServiceTest criado com 4 testes
- updates.json: v1.0.4

+++ src/main/java/my_app/db/dto/PedidoDto.java
@@ -3,7 +3,7 @@ package my_app.db.dto;
 import java.math.BigDecimal;
 
 public record PedidoDto(
-        Long clienteId,
+        Integer clienteId,
         String formaPagamento,
         BigDecimal totalLiquido,
         BigDecimal desconto,
